+++ metricbeat/docs/modules/system.asciidoc
@@ -84,9 +84,6 @@ metricbeat.modules:
   enabled: true
   period: 10s
   processes: ['.*']
-
-  # if true, exports the CPU usage in ticks, together with the percentage values
-  cpu_ticks: false
 ----
 
 [float]
