+++ metricbeat/etc/beat.yml
@@ -43,7 +43,4 @@ metricbeat.modules:
   period: 10s
   processes: ['.*']
 
-  # if true, exports the CPU usage in ticks, together with the percentage values
-  cpu_ticks: false
-
