+++ metricbeat/module/system/_meta/config.full.yml
@@ -29,3 +29,9 @@
   enabled: true
   period: 10s
   processes: ['.*']
+
+  # if true, exports the CPU usage in ticks, together with the percentage values
+  #cpu_ticks: false
+
+  # EXPERIMENTAL: cgroups can be enabled for the process metricset.
+  #cgroups: false
